move base class into WorkOS module
